requirements.txt
kubernetes
PyYAML
sentry-sdk
python-dotenv
grpcio
grpcio-tools
protobuf